@@ -1,7 +1,7 @@
 {
   "name": "dotkernel/dot-rbac-guard",
   "type": "library",
-  "description": "Various authorization guards for route/controller authorization",
+  "description": "DotKernel RBAC guards component",
   "license": "MIT",
   "authors": [
     {
@@ -10,16 +10,13 @@
     }
   ],
   "require": {
-    "php": "^5.6 || ^7.0",
+    "php": "^7.1",
     "psr/http-message": "^1.0",
     "container-interop/container-interop": "^1.1",
     "zendframework/zend-expressive-router": "^1.0",
     "zendframework/zend-servicemanager": "^3.0",
 
-    "dotkernel/dot-rbac": "0.6.x-dev",
-    "dotkernel/dot-authorization": "0.6.x-dev",
-    "dotkernel/dot-event": "0.6.x-dev",
-    "dotkernel/dot-helpers": "0.6.x-dev"
+    "dotkernel/dot-rbac": "~0.1"
   },
   "require-dev": {
     "phpunit/phpunit": "^4.8",
@@ -28,9 +25,9 @@
     "zendframework/zend-diactoros": "^1.3",
     "zendframework/zend-stdlib": "^3.0",
 
-    "dotkernel/dot-controller": "0.6.x-dev",
-    "dotkernel/dot-authentication": "0.6.x-dev",
-    "dotkernel/dot-flashmessenger": "0.6.x-dev"
+    "dotkernel/dot-controller": "~0.1",
+    "dotkernel/dot-authentication": "~0.1",
+    "dotkernel/dot-flashmessenger": "~0.1"
   },
   "autoload": {
     "psr-4": {
@@ -44,7 +41,7 @@
   },
   "extra": {
     "branch-alias": {
-      "dev-master": "0.6-dev"
+      "dev-master": "0.2-dev"
     }
   }
 }
