@@ -7,14 +7,17 @@
  * Time: 2:14 AM
  */
 
+declare(strict_types = 1);
+
 namespace Dot\Rbac\Guard\Factory;
 
+use Dot\Authentication\AuthenticationInterface;
 use Dot\Authorization\AuthorizationInterface;
-use Dot\Rbac\Guard\Listener\DefaultAuthorizationListener;
 use Dot\Rbac\Guard\Middleware\RbacGuardMiddleware;
+use Dot\Rbac\Guard\Options\RbacGuardOptions;
+use Dot\Rbac\Guard\Provider\Factory;
+use Dot\Rbac\Guard\Provider\GuardsProviderPluginManager;
 use Interop\Container\ContainerInterface;
-use Zend\EventManager\EventManager;
-use Zend\EventManager\EventManagerInterface;
 
 /**
  * Class RbacGuardMiddlewareFactory
@@ -22,22 +25,34 @@ use Zend\EventManager\EventManagerInterface;
  */
 class RbacGuardMiddlewareFactory
 {
+    use AttachAuthorizationEventListenersTrait;
+
     /**
      * @param ContainerInterface $container
+     * @param $requestedName
      * @return RbacGuardMiddleware
      */
-    public function __invoke(ContainerInterface $container)
+    public function __invoke(ContainerInterface $container, $requestedName)
     {
-        $eventManager = $container->has(EventManagerInterface::class)
-            ? $container->get(EventManagerInterface::class)
-            : new EventManager();
+        /** @var RbacGuardOptions $options */
+        $options = $container->get(RbacGuardOptions::class);
 
-        /** @var DefaultAuthorizationListener $defaultListener */
-        $defaultListener = $container->get(DefaultAuthorizationListener::class);
-        $defaultListener->attach($eventManager);
+        /** @var Factory $guardsProviderFactory */
+        $guardsProviderFactory = new Factory($container, $container->get(GuardsProviderPluginManager::class));
+        $guardsProvider = $guardsProviderFactory->create($options->getGuardsProvider());
 
-        $middleware = new RbacGuardMiddleware($container->get(AuthorizationInterface::class));
-        $middleware->setEventManager($eventManager);
+        /** @var RbacGuardMiddleware $middleware */
+        $middleware = new $requestedName(
+            $container->get(AuthorizationInterface::class),
+            $guardsProvider,
+            $options,
+            $container->has(AuthenticationInterface::class)
+                ? $container->get(AuthenticationInterface::class)
+                : null
+        );
+
+        $middleware->attach($middleware->getEventManager(), 1000);
+        $this->attachListeners($container, $middleware->getEventManager());
 
         return $middleware;
     }
