@@ -7,9 +7,11 @@
  * Time: 12:45 AM
  */
 
+declare(strict_types = 1);
+
 namespace Dot\Rbac\Guard\Provider;
 
-use Dot\Rbac\Guard\Factory\ArrayGuardsProviderFactory;
+use Dot\Rbac\Guard\Factory\GuardsProviderFactory;
 use Zend\ServiceManager\AbstractPluginManager;
 
 /**
@@ -21,6 +23,15 @@ class GuardsProviderPluginManager extends AbstractPluginManager
     protected $instanceOf = GuardsProviderInterface::class;
 
     protected $factories = [
-        ArrayGuardsProvider::class => ArrayGuardsProviderFactory::class,
+        ArrayGuardsProvider::class => GuardsProviderFactory::class,
+    ];
+
+    protected $aliases = [
+        'arrayguardsprovider' => ArrayGuardsProvider::class,
+        'arrayGuardsProvider' => ArrayGuardsProvider::class,
+        'ArrayGuardsProvider' => ArrayGuardsProvider::class,
+        'arrayguards' => ArrayGuardsProvider::class,
+        'arrayGuards' => ArrayGuardsProvider::class,
+        'ArrayGuards' => ArrayGuardsProvider::class,
     ];
 }
